Combine PR110 into this one
