+++ docs/sphinx/examples_rst/qec/circuit_level_noise.rst
@@ -76,9 +76,47 @@ Here's how to use CUDA-Q QEC to perform a circuit-level noise model experiment i
 The CUDA-Q QEC library thus provides a platform for numerical QEC experiments. The `qec.code` can be used to analyze a variety of QEC codes (both library or user provided), with a variety of decoders (both library or user provided).
 The CUDA-Q QEC library also provides tools to speed up the automation of generating noisy data and syndromes.
 
-Addtionally, here's how to use CUDA-Q QEC to construct a multi-round parity check matrix for the circuit-level noise model experiment in Python:
+Addtionally, here's how to use CUDA-Q QEC to construct a multi-round parity check matrix and a custom error correction code for the circuit-level noise model experiment in Python:
 
 .. tab:: Python
 
-   .. literalinclude:: ../../examples/qec/python/generate_parity_check_matrix.py
+   .. literalinclude:: ../../examples/qec/python/repetition_code_fine_grain_noise.py
       :language: python
+
+This example illustrates how to:
+
+1. Construct a multi-round parity check matrix – Users can extend a single-round parity check matrix across multiple rounds, 
+incorporating measurement errors to track syndrome evolution over time. This enables more accurate circuit-level noise modeling for decoders.
+
+2. Define custom error correction circuits with precise noise injection – Using `cudaq.apply_noise`, users can introduce specific error channels 
+at targeted locations within the QEC circuit. This fine-grained control allows for precise testing of how different noise sources affect logical error rates.
+
+In the previous example, we demonstrated how to introduce random X errors into each data qubit using `cudaq.apply_noise` during each round of syndrome extraction. 
+CUDA-Q allows users to inject a variety of error channels at different locations within their circuits, enabling fine-grained noise modeling. The example below showcases 
+additional ways to introduce errors into a quantum kernel:
+
+   .. code-block:: python
+
+        @cudaq.kernel
+        def inject_noise_example():
+            q = cudaq.qvector(3)
+
+            # Apply depolarization noise to the first qubit
+            cudaq.apply_noise(cudaq.DepolarizationChannel, 0.1, q[0])
+
+            # Perform gate operations
+            h(q[1])
+            x.ctrl(q[1], q[2])
+
+            # Inject a Y error into the second qubit
+            cudaq.apply_noise(cudaq.YError, 0.1, q[1])
+
+            # Apply a general Pauli noise channel to the third qubit
+            cudaq.apply_noise(cudaq.Pauli1, 0.1, 0.1, 0.1, q[2])
+
+        # Define and apply a noise model
+        noise = cudaq.NoiseModel()
+        counts = cudaq.sample(inject_noise_example, noise_model=noise)
+
+For a full list of supported noise models and their parameters, refer to the `CUDA-Q documentation <https://nvidia.github.io/cuda-quantum/latest/index.html>`_.
+
